requirements.txt
Flask
backgroundremover
Pillow
onnxruntime
torch
torchvision
ffmpeg-python
